pyproject.toml
[project]
name = "kohaLibTools"
version = "0.0.1"
description = "All Saints Koha Library Tools"
authors = [
    {name = "Stephen Gaito", email = "stephen@perceptisys.co.uk"},
]
dependencies = [
    "pyyaml>=6.0",
    "mariadb>=1.1.10",
    "nicegui>=1.3.9",
    "cryptography>=41.0.3",
    "pillow>=10.0.1",
    "python-barcode>=0.15.1",
    "fpdf2>=2.7.9",
]
requires-python = ">=3.10"
license = {text = "MIT"}

#[project.scripts]
#kohaLibTools = "kohaLibTools.cli:main"

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.pdm]
